utils: use cmark gfm for the build when running the compiler
When using the just built compiler, we need to use the build CMark
runtime as it is executing on the build. This used to work when host and
build were guaranteed to be identical (i.e. no cross-compilation). This
is now needed to support cross-compiling ARM64.

(cherry picked from commit 421c0cf890dc59a9bf2764cbe08da096ac3936e7)