[cxx-interop] Do not crash when passing `Bool` as `const T&` parameter
The type bridging logic assumed that if a value of type `Swift.Bool` is passed to a Clang function as an argument, then the type of the parameter must be a Clang built-in type (usually `_Bool`). This is not always correct. For instance, the type might be a templated const reference.

rdar://125508505
(cherry picked from commit 5b5ffde79e2573621e34d1cd0b5e3c44782735eb)