Merge pull request #73491 from tbkka/tbkka-mpe-sparebits
Calculate spare bits for multi-payload enums from first principles

This adds a getSpareBits method to all the TypeInfo classes
that returns a suitable bitmask indicating the spare bits available
in values of this type. This gives us a way to recursively explore
the type tree and build up a full spare bit mask for an arbitrary type.

We mostly have enough information to do this calculation entirely from first
principles, without requiring additional reflection information. So once this is
stable, we should remove my earlier incomplete effort to publish spare bit mask
info in the reflection data, as that adds unnecessary metadata to every binary.

TODO: Resilience forces some enums to not use spare bits even though
they otherwise would be able to.  We should have the compiler add a single
bit to the reflectio data indicating whether or not spare bits were used
and then rely on this code to actually compute the spare bits.

This doubtless still has plenty of holes, but seems sufficient
to handle a few basic enum types, including the stdlib DecodingError
which was used as an example to work out some key issues.

Resolves rdar://126563813