Adding stage-0 toolchain build preset
This is a preset for building a stage-0 compiler toolchain that doesn't
require an existing Swift installation. The resulting compiler should
not be used outside of bringing up a stage-1 compiler, which in turn
should build a stage 2 compiler. That stage 2 compiler should run and
test cleanly. The stage-0 compiler will definitely fail its tests, and
the stage-1 compiler may or may not depending on the weather. The
stage-2 compiler is the compiler toolchain that should be shippable.