[Macros] Fix source location for declaration syntax of attached macros
Use the correct source file.
Previously, the location of declaration macros were sent to executable
plugins as if the were in the source file of the attribute. This was
problematic when the attribute is synsthesized by a macro. When source
location were used in the plugin, for example, as a diagnostic location,
it ended up with an unknown location.

(cherry picked from commit 921555ab38bd8ede534b57fd2f071a245eae47f9)