[region-isolation] Rely on our classification of lookthrough or not to find underlying tracked object values rather than delegating to other analysis helpers.
There is no guarantee that these other helpers properly model lookthrough as our
model does. This ensures that this routine is always in sync with how we define
lookthrough in our model.

The problem with the old approach can be seen in how we handled move_value. The
model and the later code knew correctly that they should not look through
move_value that is marked as [var_decl]. But this other analysis code did not.
This with the tree today should not have any impact. But with the fix I am doing
now (fixing nonisolated(unsafe)) and later isolation history this will become a
problem.

(cherry picked from commit b66cfccef6b441ab0002adb53a8af648c2032120)