Add target dependencies to the package manifest via Fix-Its
When a module name for a given import declaration cannot be resolved,
consult the "available modules" file provided by the Swift package
manager to determine whether a module by that name is available
something in the package or any of its dependencies.

If so, add a note with a Fix-It to the "no such module" diagnostic
that edits the package manifest to introduce the missing target
dependencies, e.g.,

    .product(name: "Testing", package: "swift-testing")

Accepting the Fix-It should update the package manifest appropriately,
so that building again will succeed in finding the appropriate module.

To support this functionality, there are three new frontend parameters:

* `-package-manifest <path>`: provides the path to the package
manifest file containing the target being built
* `-package-target-name <name>`: provides the name of the target being
built within that package manifest file
* `-package-available-modules <path>`: provides the path to a JSON
file containing the set of modules that are known to the package,
which includes both modules declared within that package as well as
any modules provided by the products of imported packages.

The Swift Package Manager is responsible for providing this
command-line parameters.