[cxx-interop] Avoid treating some Obj-C types as foreign reference types
This makes sure we don't apply logic that is specific to C++ reference types to Objective-C types.

Previously we were mistakenly treating some Objective-C types as foreign reference types. This meant that IRGen would try to emit calls to custom lifetime operations. This should not happen for non-C++ types.

rdar://128447046