Sema: Ban unsound convenience initializer inheritance inside a module
We already prevent the user from subclassing a class like this from
another module,

    open class C {
      private init() {}
      public convenience init(x: Int) { self.init() }
    }

However, we cannot allow this within a module also:

    class C {
      private init() {}
      convenience init(x: Int) { self.init() }
    }

    class D: C {}

Today, we don't diagnose anything, and we also don't synthesize the
override, so we just miscompile instead.

We can't synthesize the override though, because this would then
allow calling the superclass's private init from the subclass, which
violates access control.

So the only option is to ban this.

Fixes rdar://68220237.
Fixes https://github.com/apple/swift/issues/61508.