[ast] Always emit compiler(>= 5.3) when emitting guards for suppressable features to guard against parser errors.
The code here was assuming that if we already emitted a compiler guard for
non-Suppressable features, we could avoid doing it for suppressable
features. The problem with this is that compiler() does more than just check for
compiler versions... it also tells the compiler that parser errors in the if
block should be ignored when if evaluates to false.

rdar://129045783
(cherry picked from commit 521a70f4167ddcbd820aa10186b5fa0d6670ac40)